halauth: add authentication object
HalAuth is a new object intended to sit in Halibot core, and handles permission
objects.

Authentication objects currently take the form of a triple:
(RI, identity, Permission String)
...which binds a particular permission to a identity and RI.

An example would be the following:
(irc/##halibot, richteer, CanReloadModules)
...which essentially says that the user "richteer" in the IRC channel "##halibot"
is allowed to operate anything that requests the "CanReloadModules" permission.

Any of the three can be wildcarded, including all of them at once. Currently,
the wildcard indicator is an asterisk ("*"). This indicates that ANY of that
particular field matches, so allowing all permissions to a particular user in
a particular channel (RI) would look like this:
(*, richteer, *)

The current API on the HalAuth module includes three methods:
 - grantPermission(ri, identity, perm) -> add a perm triple to the list
 - revokePermission(ri, identity, perm) -> remove an EXACT MATCHING triple
 - hasPermission(ri, identity, perm) -> check if a triple matches one in the list